requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
git+https://github.com/mikeckennedy/fastapi-chameleon
pyyaml